netfree/backend/requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
sqlalchemy==2.0.31
pydantic==2.7.4
httpx==0.27.0
python-dotenv==1.0.1
psycopg2-binary==2.9.9
alembic==1.13.2
